+++ packages/common/src/@types/options/resolved.ts
@@ -1,8 +1,9 @@
-import type { Format, Omit } from "ts-vista";
+import type { Format, Omit, Partial } from "ts-vista";
 
 import type { CompleteOptions } from "#/@types/options/complete";
 import type {
     CompleteBuildHandlerOptions,
+    CompleteBuildPublicOptions,
     CompleteBuildServerOptions,
 } from "#/@types/options/complete/build";
 import type { CompleteDevOptions } from "#/@types/options/complete/dev";
@@ -12,6 +13,13 @@ import type { HttpsOptions } from "#/@types/options/default";
  * Resolved HTTPS server options.
  */
 type ResolvedHttpsOptions = HttpsOptions;
+
+/**
+ * Resolved public directory options.
+ */
+type ResolvedBuildPublicOptions = Format<
+    Partial<CompleteBuildPublicOptions, "to">
+>;
 
 /**
  * Resolved development server options.
@@ -29,18 +37,29 @@ type ResolvedDevOptions = Format<
  * Resolved build server options.
  */
 type ResolvedBuildServerOptions = Format<
-    Omit<CompleteBuildServerOptions, "https"> & {
+    Omit<CompleteBuildServerOptions, "https" | "public"> & {
         /**
          * HTTPS server options.
          */
         https?: HttpsOptions;
+        /**
+         * Public directory options.
+         */
+        public: ResolvedBuildPublicOptions;
     }
 >;
 
 /**
  * Resolved build handler options.
  */
-type ResolvedBuildHandlerOptions = Format<CompleteBuildHandlerOptions>;
+type ResolvedBuildHandlerOptions = Format<
+    Omit<CompleteBuildHandlerOptions, "public"> & {
+        /**
+         * Public directory options.
+         */
+        public: ResolvedBuildPublicOptions;
+    }
+>;
 
 /**
  * Resolved build options.
@@ -68,6 +87,7 @@ type ResolvedOptions = Format<
 export type {
     ResolvedBuildHandlerOptions,
     ResolvedBuildOptions,
+    ResolvedBuildPublicOptions,
     ResolvedBuildServerOptions,
     ResolvedDevOptions,
     ResolvedHttpsOptions,